requirements.txt
ipython==7.7.0
jupyter==1.0.0
requests==2.22.0
ipdb
django
djangorestframework
markdown
django-filter
